multi-domain-clean/requirements.txt
Flask>=2.3.0
pymysql>=1.1.0
cryptography>=41.0.0
requests>=2.31.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
